notebooks/bonus/requirements.txt
# W4 Voice Bonus - Additional Requirements
# Install with: pip install -r requirements.txt

# Core (already in main requirements.txt)
openai>=1.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0

# Realtime API (W4_realtime_voice_agent.py)
websockets>=12.0
sounddevice>=0.4.6
numpy>=1.24.0
